+++ backend/cwa_lib/pydantic_schemas/api_groups.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class GroupCreate(BaseModel):
-    group_id: int
-    group_name: str